Small bug fixes, flow distribution working
Path's info still screwing up the "master connection" from the T, but flow distribution could be implemented, just maybe not in the best way. Cylinder got to have some checks for not changing values so constantly as the motor, since it doesn't keep moving when flow stops like the motor.